+++ common/environment.go
@@ -32,6 +32,9 @@ type EnvironmentVariable struct {
 }
 
 // This array needs to be updated when a new public environment variable is added
+// Things are here, rather than in command line parameters for one of two reasons:
+// 1. They are optional and obscure (e.g. performance tuning parameters) or
+// 2. They are authentication secrets, which we do not accept on the command line
 var VisibleEnvironmentVariables = []EnvironmentVariable{
 	EEnvironmentVariable.ConcurrencyValue(),
 	EEnvironmentVariable.TransferInitiationPoolSize(),